Do a bitwise XOR with a 4 digit PIN and the shared secret before encoding the shared secret into a QR code.
With a bitwise XOR we ensure that the shared secret is protected untill
it arrives in the client application. XOR is a good cryptographic
operation for this purpose, because if you XOR a PIN with an XORed
shared secret you do not know if it is valid. Only the server has can
check if the shared secret is valid.